backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
sqlalchemy>=2.0.0
python-jose[cryptography]>=3.3.0
pyotp>=2.9.0
bcrypt>=4.2.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
